# docs/swagger.yaml
definitions:
...
  entities.Currency:
    properties:
      additional_alias_rate:
        type: number
      alias:
        type: string
      base_currency:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  entities.CurrencyAttributes:
    properties:
      multiplier:
        type: integer
      synonym:
        type: string
      title:
        type: string
    type: object
  entities.CurrencyExchange:
    properties:
      created_at:
        type: string
      from:
        type: string
      rate:
        type: number
      to:
        type: string
    type: object
  entities.CurrencyMultiplier:
    properties:
      multiplier:
        type: integer
      organization_pair_id:
        type: string
      provider_integrator_pair:
        $ref: '#/definitions/entities.ProviderIntegratorPair'
      synonym:
        type: string
      title:
        type: string
    type: object
...
  entities.Organization:
    properties:
      api_key:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: integer
      type:
        type: string
    type: object
...
  entities.ProviderIntegratorPair:
    properties:
      id:
        type: string
      integrator:
        $ref: '#/definitions/entities.Organization'
      integrator_id:
        type: string
      provider:
        $ref: '#/definitions/entities.Organization'
      provider_id:
        type: string
    type: object
...
  requests.Currency:
    properties:
      additional_alias_rate:
        minimum: 1
        type: number
      alias:
        minLength: 1
        type: string
      base_currency:
        type: string
      title:
        minLength: 1
        type: string
      type:
        minLength: 1
        type: string
    required:
    - additional_alias_rate
    type: object
  requests.CurrencyExchangeRequest:
    properties:
      from:
        type: string
      rate:
        type: number
    required:
    - from
    - rate
    type: object
  requests.CurrencyMultiplier:
    properties:
      multiplier:
        type: integer
      organization_pair_id:
        type: string
      synonym:
        minLength: 1
        type: string
      title:
        minLength: 1
        type: string
    type: object
  requests.CurrencyMultiplierIdentify:
    properties:
      organization_pair_id:
        type: string
      title:
        minLength: 1
        type: string
    type: object
...
  requests.CurrencySearchRequest:
    properties:
      organization_pair_id:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      meta: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/currency:
    delete:
      consumes:
      - application/json
      description: Delete currency.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: alias
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: Delete currency.
      tags:
      - currency
    get:
      consumes:
      - application/json
      description: Get all currencies` information.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Currency'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get all currencies.
      tags:
      - currency
    post:
      consumes:
      - application/json
      description: Create new currency.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CreateCurrency
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.Currency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Currency'
              type: object
      security:
      - X-Authenticate: []
      summary: Create new currency.
      tags:
      - currency
  /api/currency/{alias}:
    get:
      consumes:
      - application/json
      description: Get currency information.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: alias
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Currency'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get currency.
      tags:
      - currency
  /api/currency/{alias}/exchange:
    delete:
      consumes:
      - application/json
      description: Delete an existing exchange rate between the specified "from" currency
        and the "to" currency (alias).
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Alias of the 'to' currency.
        in: path
        name: alias
        required: true
        type: string
      - description: CurrencyExchangeRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencyExchangeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - X-Authenticate: []
      summary: Delete a currency exchange rate.
      tags:
      - currency
    get:
      consumes:
      - application/json
      description: Get currency exchange information.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: alias
        in: path
        name: title
        required: true
        type: string
      - description: 'ordering format: column decs/acs'
        in: query
        name: order
        type: string
      - description: spins per page
        in: query
        name: limit
        required: true
        type: string
      - description: asking page
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.CurrencyExchange'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get currency exchange.
      tags:
      - currency
    post:
      consumes:
      - application/json
      description: Add a new exchange rate between the specified "from" currency and
        the "to" currency (alias).
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Alias of the 'to' currency.
        in: path
        name: alias
        required: true
        type: string
      - description: CurrencyExchangeRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencyExchangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.CurrencyExchange'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - X-Authenticate: []
      summary: Add a new currency exchange rate.
      tags:
      - currency
  /api/currency/multiplier:
    delete:
      consumes:
      - application/json
      description: Delete currency and bound multiplier by provider_id and integrator_id.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CurrencyMultiplierIdentify
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencyMultiplierIdentify'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: Delete currency for specific provider and integrator.
      tags:
      - currency
    post:
      consumes:
      - application/json
      description: Create new currency and bound multiplier by provider_id and integrator_id.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CreateCurrencyMultiplier
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencyMultiplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.CurrencyMultiplier'
              type: object
      security:
      - X-Authenticate: []
      summary: Create new currency for specific provider and integrator.
      tags:
      - currency
    put:
      consumes:
      - application/json
      description: Updates currency and bound multiplier by provider_id and integrator_id.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CreateCurrencyMultiplier
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencyMultiplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.CurrencyMultiplier'
              type: object
      security:
      - X-Authenticate: []
      summary: Updates currency for specific provider and integrator.
      tags:
      - currency
  /api/currency/multiplier/download:
    post:
      consumes:
      - application/json
      description: This endpoint allows you to search for currency data based on filters
        and download it as an Excel file.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Search Criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencySearchRequest'
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file with currency data
          schema:
            type: file
      security:
      - X-Authenticate: []
      summary: Download currency data as an Excel file
      tags:
      - currency
  /api/currency/multiplier/get:
    post:
      consumes:
      - application/json
      description: Get currency multiplier information.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CurrencyMultiplierIdentify
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencyMultiplierIdentify'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.CurrencyMultiplier'
              type: object
      security:
      - X-Authenticate: []
      summary: Get currency multiplier.
      tags:
      - currency
  /api/currency/multiplier/search:
    post:
      consumes:
      - application/json
      description: |-
        Filtering can be done by organization_pair_id or you can send an empty request body
        and get all currency.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CurrencySearch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CurrencySearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.CurrencyMultiplier'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get all currency by filter.
      tags:
      - currency
  /api/currency/multiplier/upload:
    post:
      consumes:
      - multipart/form-data
      description: This endpoint allows you to upload currency data from an Excel
        file and save it to the database.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Organization Pair ID (UUID format)
        in: formData
        name: organization_pair_id
        required: true
        type: string
      - description: Excel file containing currency data
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Uploaded currency data
          schema:
            items:
              $ref: '#/definitions/entities.CurrencyAttributes'
            type: array
      security:
      - X-Authenticate: []
      summary: Upload currency data from an Excel file
      tags:
      - currency
...
swagger: "2.0"